Rewrite LCDSocketPoller to use blocking I/O instead of polling
This made the shutdown procedure a little more interesting since
BufferedReader's read(Line) cannot easily be interrupted. One possible
solution is to use shutdownInput() on the socket to provoke an End of
Stream [1], which is what is done here. The other way is to just close
the socket, causing an IOException. I opted not to do that since I would
have to suppress all IOExceptions in basically the entire run() method
in LCDSocketPoller

[1]: https://stackoverflow.com/questions/3595926/how-to-interrupt-bufferedreaders-readline