Communicate LCDd readiness with a Semaphore instead of polling
Access to _lastLine is now properly synchronized. It is a bit ugly to
use a field just to send back a status line once, but it seems the
easiest way in the current implementation. It is also not customary to
increment a semaphore more times than you plan on decrementing it, but
it's not strictly illegal either.

LCD::connect now only tries to connect once. Connection issues should be
rare, and it is up to the application what to do if the first attempt
should fail.